chore: Rename QuotaCreatedForTierDomainEventHandler to TierQuotaDefinitionCreatedDomainEventHandler
