+++ Modules/Quotas/src/Quotas.Application/Extensions/IEventBusExtensions.cs
@@ -2,13 +2,13 @@ using Backbone.BuildingBlocks.Application.Abstractions.Infrastructure.EventBus;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.FileUploaded;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.IdentityCreated;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.MessageCreated;
-using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.QuotaCreatedForTier;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.RelationshipChangeCompleted;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.RelationshipChangeCreated;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.RelationshipTemplateCreated;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.TierCreated;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.TierDeleted;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.TierOfIdentityChanged;
+using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.TierQuotaDefinitionCreated;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.TierQuotaDefinitionDeleted;
 using Backbone.Modules.Quotas.Application.DomainEvents.Incoming.TokenCreated;
 using Backbone.Modules.Quotas.Domain.DomainEvents.Incoming.FileUploaded;
@@ -37,7 +37,7 @@ public static class IEventBusExtensions
         eventBus.Subscribe<IdentityCreatedDomainEvent, IdentityCreatedDomainEventHandler>();
         eventBus.Subscribe<TierCreatedDomainEvent, TierCreatedDomainEventHandler>();
         eventBus.Subscribe<TierDeletedDomainEvent, TierDeletedDomainEventHandler>();
-        eventBus.Subscribe<TierQuotaDefinitionCreatedDomainEvent, QuotaCreatedForTierDomainEventHandler>();
+        eventBus.Subscribe<TierQuotaDefinitionCreatedDomainEvent, TierQuotaDefinitionCreatedDomainEventHandler>();
         eventBus.Subscribe<MessageCreatedDomainEvent, MessageCreatedDomainEventHandler>();
         eventBus.Subscribe<TierQuotaDefinitionDeletedDomainEvent, TierQuotaDefinitionDeletedDomainEventHandler>();
         eventBus.Subscribe<FileUploadedDomainEvent, FileUploadedDomainEventHandler>();
